server/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
dj-database-url==3.0.1
django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-unfold==0.62.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
docx==0.2.4
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
jinja2==3.1.6
langcodes==3.5.0
language-data==1.3.0
lazy-loader==0.4
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
preshed==3.0.10
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.9.0
pymupdf==1.26.3
pypdf2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-docx==1.2.0
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.32.4
rich==14.1.0
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
pinecone 
sentence-transformers 
rank-bm25 
langchain
pandas
nltk